+++ cli/src/main/scala/bondlink/FindUnusedCli.scala
@@ -1,4 +1,4 @@
-package bondlink
+package bl.unused
 
 import cats.syntax.either.*
 import cats.syntax.foldable.*
